[CFN-27] Working towards a more extensible DBInstance create flow. Static methods and error rules have been pulled out of the base handler class; The instance create factory is now constructed by factory according to the model.
