+++ aws-rds-cfn-common/src/main/java/software/amazon/rds/ProgressStep.java
@@ -1,0 +1,4 @@
+package software.amazon.rds;
+
+public class ProgressStep {
+}
