+++ aws-rds-dbinstance/src/main/java/software/amazon/rds/dbinstance/common/create/FromSnapshot.java
@@ -1,0 +1,14 @@
+package software.amazon.rds.dbinstance.common.create;
+
+import org.apache.commons.lang3.NotImplementedException;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.rds.dbinstance.CallbackContext;
+import software.amazon.rds.dbinstance.ResourceModel;
+
+public class FromSnapshot implements DBInstanceFactory {
+
+    @Override
+    public ProgressEvent<ResourceModel, CallbackContext> create(ProgressEvent<ResourceModel, CallbackContext> input) {
+        throw new NotImplementedException();
+    }
+}
